vendorize the gnmi cache
We'll have to tweak the gnmi cache behavior and the easiest way
to do so is to vendorize it. We need to allow emitting periodic
metrics even if there aren't any changes, while the upstream
gnmi cache will drop such notifications.

Alternatives:

* vendorize the entire gnmi library - not necessary, more difficult
to maintain
* bypass the cache when hitting stale update errors or completely
disable the cache - might have unforeseen consequences