Improve request target wildcards
Right now, each target is supposed to have an associated request,
while the requests have either a backreference to a target or
a "*" target (default requests).

We need a more flexible way of defining gnmi requests, so we're
making the following changes:

* targets no longer have to explicitly reference a request
* the requests can reference multiple targets by using glob patterns
  such as ce*, tor*, etc.

Since the gnmi library still requires one request per target, we're
going to aggregate the requests that gnmi-gw receives.

For now, we're only covering the zookeeper target loader.