Fix misleading log message
We're logging "No matching target for request" despite actually
finding a target match. The reason is that the "found" variable
is reused by the same code block but for a different purpose.

We'll fix this issue by adding another variable called "foundSub",
which is going to be used when searching for existing subscriptions.