Avoid deadlock when reconnecting
The gnmi cache gets reset when removing or recreating connections.
Any cached notification will be pushed to the exporters.

The issue is that ZookeeperConnectionManager.Forwardable loops
over the connections, tries to acquire the connection lock and
enters a deadlock.

A simple workaround would be to just invoke the disconnect/reconnect
calls asynchronously. Worth mentioning that "connect" call is already
async.

At the same time, we can restore the old reconnect implementation,
which is simply closing the underlying connection, which in turn will
trigger a reconnect attempt.