Allow associating metadata with gnmi paths
Our exporters need additional information based on the gnmi path
(e.g. whether to handle the data as a metric or as an event, the
log namespace, etc).

We'll allow defining arbitrary metadata for groups of gnmi paths,
which will be stored in Zookeeper.

Child gnmi nodes inherit their parent metadata. Let's take the
following example:

	path: /a
	metadata: {"a": "parent", "b": "b"}

	path: /a/child
	metadata: {"a": "child"}

The resulting metadata for "/a/child" will be {"a": "child", "b": "b"}.