statsd: use buffered client
The buffered statsd client tries to push multiple notifications
at a time in order to decrease overhead.

It's using a configurable delay, by default 300ms. If another
notification is received in that time span, the two notifications
are merged.

The notification is submitted to statsd when either the delay
interval or maximum buffer size are exceeded. By default, the
buffer is capped to ~1400 bytes in order to avoid fragmentation.

We've tested the two udp statsd clients using 150k requests.
The buffered one decreased the total duration from 4.1s to 1.5s,
keeping the average latency below 1ms. The cpu usage is also
expected to be significantly lower.